+++ apps/api/routes/employees.ts
@@ -1,44 +1,37 @@
-import { FastifyPluginOptions, FastifyInstance, FastifyRequest } from "fastify";
-import { IdParamsSchema, IdParamsType } from "../schema/id";
+import { FastifyPluginOptions, FastifyInstance } from "fastify";
+import { IdParamsSchema, IdParamsType } from "database/schema/id";
 import {
   EmployeeCreateSchema,
   EmployeeCreateBody,
   EmployeeUpdateBody,
   EmployeeUpdateSchema,
-} from "../schema/employees";
+} from "database/schema/employees";
 // import Employees from "../models/EmployeeModel";
-import { Employee, hashPassword } from "database";
-import type { Extended } from "helpers";
-type ExtendedEmployee = Extended<Employee>;
-function excludePassword(
-  employees: ExtendedEmployee[] | ExtendedEmployee | null
-) {
-  if (!employees) return null;
-  if (Array.isArray(employees)) {
-    return employees.map((employee) => {
-      const { password, ...rest } = employee;
-      return rest;
-    });
-  } else {
-    const { password, ...rest } = employees;
-    return rest;
-  }
-}
+import { hashPassword } from "database";
+import { excludePassword } from "database";
+
 async function routes(fastify: FastifyInstance, options: FastifyPluginOptions) {
-  fastify.get("/", async function (req, reply) {
-    const employees = await fastify.prisma.employee.findMany({
-      include: {
-        merchant: true,
-      },
-      orderBy: {
-        id: "asc",
-      },
-    });
-    return excludePassword(employees);
-  });
+  fastify.get(
+    "/",
+    { onRequest: [fastify.auth.verifySystemAdmin] },
+    async function (req, reply) {
+      const employees = await fastify.prisma.employee.findMany({
+        include: {
+          merchant: true,
+        },
+        orderBy: {
+          createdAt: "desc",
+        },
+      });
+      return excludePassword(employees);
+    }
+  );
   fastify.get<{ Params: IdParamsType }>(
     "/:id",
-    { schema: { params: IdParamsSchema } },
+    {
+      onRequest: [fastify.auth.verifyEmployee],
+      schema: { params: IdParamsSchema },
+    },
     async function (req, reply) {
       const employee = await fastify.prisma.employee.findUnique({
         where: {
@@ -48,17 +41,33 @@ async function routes(fastify: FastifyInstance, options: FastifyPluginOptions) {
           merchant: true,
         },
       });
+      if (!employee) {
+        reply.statusCode = 404;
+        throw new Error("Employee not found");
+      }
+      fastify.auth.verifyHasMerchantPermission(
+        req,
+        reply,
+        employee.merchantId as number
+      );
       return excludePassword(employee);
     }
   );
   fastify.post<{ Body: EmployeeCreateBody }>(
     "/",
-    { schema: { body: EmployeeCreateSchema } },
+    {
+      onRequest: [fastify.auth.verifyEmployee],
+      schema: { body: EmployeeCreateSchema },
+    },
     async function (req, reply) {
-      const { password, ...rest } = req.body;
+      const { password, merchantId, systemAdmin, ...rest } = req.body;
+      fastify.auth.verifyHasMerchantPermission(req, reply, merchantId);
       const employee = await fastify.prisma.employee.create({
         data: {
           ...rest,
+          merchantId:
+            req.employee?.systemAdmin && systemAdmin ? null : merchantId,
+          systemAdmin: req.employee?.systemAdmin && systemAdmin ? true : false,
           password: hashPassword(password),
         },
       });
@@ -67,15 +76,35 @@ async function routes(fastify: FastifyInstance, options: FastifyPluginOptions) {
   );
   fastify.put<{ Body: EmployeeUpdateBody; Params: IdParamsType }>(
     "/:id",
-    { schema: { body: EmployeeUpdateSchema, params: IdParamsSchema } },
+    {
+      onRequest: [fastify.auth.verifyEmployee],
+      schema: { body: EmployeeUpdateSchema, params: IdParamsSchema },
+    },
     async function (req, reply) {
-      const { password, ...rest } = req.body;
+      const currentEmployee = await fastify.prisma.employee.findUnique({
+        where: {
+          id: req.params.id,
+        },
+      });
+      if (!currentEmployee) {
+        reply.statusCode = 404;
+        throw new Error("Employee not found");
+      }
+      fastify.auth.verifyHasMerchantPermission(
+        req,
+        reply,
+        currentEmployee.merchantId
+      );
+      const { password, systemAdmin, merchantId, ...rest } = req.body;
       const employee = await fastify.prisma.employee.update({
         where: {
           id: req.params.id,
         },
         data: {
           ...rest,
+          systemAdmin: req.employee?.systemAdmin && systemAdmin ? true : false,
+          merchantId:
+            req.employee?.systemAdmin && systemAdmin ? null : merchantId,
           ...(password
             ? {
                 password: hashPassword(password),
@@ -88,8 +117,25 @@ async function routes(fastify: FastifyInstance, options: FastifyPluginOptions) {
   );
   fastify.delete<{ Params: IdParamsType }>(
     "/:id",
-    { schema: { params: IdParamsSchema } },
+    {
+      onRequest: [fastify.auth.verifyEmployee],
+      schema: { params: IdParamsSchema },
+    },
     async function (req, reply) {
+      const currentEmployee = await fastify.prisma.employee.findUnique({
+        where: {
+          id: req.params.id,
+        },
+      });
+      if (!currentEmployee) {
+        reply.statusCode = 404;
+        throw new Error("Employee not found");
+      }
+      fastify.auth.verifyHasMerchantPermission(
+        req,
+        reply,
+        currentEmployee.merchantId
+      );
       await fastify.prisma.employee.delete({
         where: {
           id: req.params.id,
